+++ src/Behat/Testwork/Exception/Stringer/PHPUnitExceptionStringer.php
@@ -11,8 +11,6 @@
 namespace Behat\Testwork\Exception\Stringer;
 
 use Exception;
-use PHPUnit_Framework_Exception;
-use PHPUnit_Framework_TestFailure;
 
 /**
  * Strings PHPUnit assertion exceptions.
@@ -28,7 +26,8 @@ final class PHPUnitExceptionStringer implements ExceptionStringer
      */
     public function supportsException(Exception $exception)
     {
-        return $exception instanceof PHPUnit_Framework_Exception;
+        return $exception instanceof \PHPUnit_Framework_Exception
+            || $exception instanceof \PHPUnit\Framework\Exception;
     }
 
     /**
@@ -36,9 +35,13 @@ final class PHPUnitExceptionStringer implements ExceptionStringer
      */
     public function stringException(Exception $exception, $verbosity)
     {
+        if (!class_exists('PHPUnit\\Framework\\TestFailure')) {
+            return trim(\PHPUnit_Framework_TestFailure::exceptionToString($exception));
+        }
+
         // PHPUnit assertion exceptions do not include expected / observed info in their
         // messages, but expect the test listeners to format that info like the following
         // (see e.g. PHPUnit_TextUI_ResultPrinter::printDefectTrace)
-        return trim(PHPUnit_Framework_TestFailure::exceptionToString($exception));
+        return trim(\PHPUnit\Framework\TestFailure::exceptionToString($exception));
     }
 }