+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
   * [#1054](https://github.com/Behat/Behat/pull/1054): [PSR-11](https://github.com/php-fig/fig-standards/blob/master/accepted/PSR-11-container.md)
     support for helper containers.
+  * Support for modern PHPUnit.
 
 ### Fixed
   * [#1056](https://github.com/Behat/Behat/pull/1056): Make Gherkin aware of the